ailyzer_for_git/employer/backend/app/pyproject.toml
[project]
name = "employer-page"
version = "0"
description = ""
authors = [
    {name = "Mikhail",email = "misabaranov8@gmail.com"}
]
license = {text = "ARPL"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn[standart] (>=0.38.0,<0.39.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
